Fixed possible overflow in Sort() on vector counts, type was int instead of size_t but counted values could increment size_t times

+++ lsd.h
@@ -80,7 +80,7 @@ namespace fsu
 	void LSDSort::Sort(fsu::Vector<fsu::String>& stringList, size_t max, size_t length)
 	{
 		fsu::Vector<fsu::String> temp(length);
-		fsu::Vector<int> counts;
+		fsu::Vector<size_t> counts;
 		size_t i;
 
 		for(i = max - 1; i < max; --i)
